Tidy SierraType
