+++ sierra2mlir/src/compiler.rs
@@ -14,7 +14,7 @@ use melior_next::{
 use std::{cmp::Ordering, collections::HashMap};
 
 use crate::{
-    libfuncs::lib_func_def::SierraLibFunc, types::DEFAULT_PRIME,
+    libfuncs::lib_func_def::SierraLibFunc, sierra_type::SierraType, types::DEFAULT_PRIME,
     userfuncs::user_func_def::UserFuncDef,
 };
 
@@ -24,164 +24,6 @@ pub struct Compiler<'ctx> {
     pub module: Module<'ctx>,
     pub main_print: bool,
     pub print_fd: i32,
-}
-
-// We represent a struct as a contiguous list of types, like sierra does, for now.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum SierraType<'ctx> {
-    Simple(Type<'ctx>),
-    Struct {
-        ty: Type<'ctx>,
-        field_types: Vec<Self>,
-    },
-    Enum {
-        ty: Type<'ctx>,
-        tag_type: Type<'ctx>,
-        storage_bytes_len: u32,
-        storage_type: Type<'ctx>, // the array
-        variants_types: Vec<Self>,
-    },
-    Array {
-        /// (u32, u32, ptr)
-        ///
-        /// (length, capacity, data)
-        ty: Type<'ctx>,
-        len_type: Type<'ctx>, // type of length and capacity: u32
-        element_type: Box<Self>,
-    },
-}
-
-impl<'ctx> SierraType<'ctx> {
-    pub fn get_width(&self) -> u32 {
-        match self {
-            SierraType::Simple(ty) => ty.get_width().unwrap_or(0),
-            SierraType::Struct { ty: _, field_types } => {
-                let mut width = 0;
-                for ty in field_types {
-                    width += ty.get_width();
-                }
-                width
-            }
-            SierraType::Enum {
-                ty: _,
-                tag_type,
-                storage_bytes_len: storage_type_len,
-                storage_type: _,
-                variants_types: _,
-            } => tag_type.get_width().unwrap() + (storage_type_len * 8),
-            SierraType::Array { ty: _, len_type, element_type: _ } => {
-                // 64 is the pointer size, assuming here
-                // TODO: find a better way to find the pointer size? it would require getting the context here
-                len_type.get_width().unwrap() * 2 + 64
-            }
-        }
-    }
-
-    pub fn get_felt_representation_width(&self) -> usize {
-        match self {
-            SierraType::Simple(_) => 1,
-            SierraType::Struct { ty: _, field_types } => {
-                field_types.iter().map(Self::get_felt_representation_width).sum()
-            }
-            SierraType::Enum {
-                ty: _,
-                tag_type: _,
-                storage_bytes_len: _,
-                storage_type: _,
-                variants_types,
-            } => {
-                1 + variants_types
-                    .iter()
-                    .map(Self::get_felt_representation_width)
-                    .max()
-                    .unwrap_or(0)
-            }
-            SierraType::Array { .. } => 2,
-        }
-    }
-
-    /// Returns the MLIR type of this sierra type
-    pub const fn get_type(&self) -> Type {
-        match self {
-            Self::Simple(ty) => *ty,
-            Self::Struct { ty, field_types: _ } => *ty,
-            Self::Enum {
-                ty,
-                tag_type: _,
-                storage_bytes_len: _,
-                storage_type: _,
-                variants_types: _,
-            } => *ty,
-            Self::Array { ty, .. } => *ty,
-        }
-    }
-
-    pub fn get_type_location(&self, context: &'ctx Context) -> (Type<'ctx>, Location<'ctx>) {
-        (
-            match self {
-                Self::Simple(ty) => *ty,
-                Self::Struct { ty, field_types: _ } => *ty,
-                Self::Enum {
-                    ty,
-                    tag_type: _,
-                    storage_bytes_len: _,
-                    storage_type: _,
-                    variants_types: _,
-                } => *ty,
-                Self::Array { ty, .. } => *ty,
-            },
-            Location::unknown(context),
-        )
-    }
-
-    /// Returns a vec of field types if this is a struct type.
-    pub fn get_field_types(&self) -> Option<Vec<Type>> {
-        match self {
-            SierraType::Simple(_) => None,
-            SierraType::Struct { ty: _, field_types } => {
-                Some(field_types.iter().map(|x| x.get_type()).collect_vec())
-            }
-            SierraType::Enum {
-                ty: _,
-                tag_type: _,
-                storage_bytes_len: _,
-                storage_type: _,
-                variants_types,
-            } => Some(variants_types.iter().map(|x| x.get_type()).collect()),
-            SierraType::Array { .. } => None,
-        }
-    }
-
-    pub fn get_field_sierra_types(&self) -> Option<&[Self]> {
-        match self {
-            SierraType::Simple(_) => None,
-            SierraType::Struct { ty: _, field_types } => Some(field_types),
-            SierraType::Enum {
-                ty: _,
-                tag_type: _,
-                storage_bytes_len: _,
-                storage_type: _,
-                variants_types,
-            } => Some(variants_types),
-            SierraType::Array { .. } => None,
-        }
-    }
-
-    /// gets the sierra array type for the given element
-    pub fn get_array_type<'c>(
-        compiler: &'c Compiler<'c>,
-        element: SierraType<'c>,
-    ) -> SierraType<'c> {
-        SierraType::Array {
-            ty: compiler.struct_type(&[
-                compiler.u32_type(),
-                compiler.u32_type(),
-                compiler.llvm_ptr_type(),
-            ]),
-            len_type: compiler.u32_type(),
-            element_type: Box::new(element),
-        }
-    }
 }
 
 /// Types, functions, etc storage.
