+++ sierra2mlir/src/types/mod.rs
@@ -4,7 +4,10 @@ use itertools::Itertools;
 use melior_next::ir::Type;
 use tracing::debug;
 
-use crate::compiler::{Compiler, SierraType, Storage};
+use crate::{
+    compiler::{Compiler, Storage},
+    sierra_type::SierraType,
+};
 
 pub const DEFAULT_PRIME: &str =
     "3618502788666131213697322783095070105623107215331596699973092056135872020481";
